feat: Fazendo o filtro funcionar para clientes

+++ src/app/pages/clients/client-card/client-card.component.ts
@@ -1,5 +1,5 @@
-import { DatePipe } from '@angular/common';
-import { Component, EventEmitter, Input, Output, inject, signal } from '@angular/core';
+import { DatePipe, NgIf } from '@angular/common';
+import { Component, EventEmitter, Input, Output, computed, inject, signal } from '@angular/core';
 import { MatButtonModule } from '@angular/material/button';
 import { MatCardModule } from '@angular/material/card';
 import { MatAccordion, MatExpansionModule } from '@angular/material/expansion';
@@ -13,6 +13,8 @@ import { DialogModule } from 'primeng/dialog';
 import { User } from '../../../models/user';
 import { ClientService } from '../../../services/client.service';
 import ClientsComponent from '../clients.component';
+import BaseStoreComponent from '../../../base/base-store.component';
+import { selectUserInfo } from '../../../store/reducers/user.reducer';
 
 @Component({
   selector: 'app-client-card',
@@ -28,13 +30,14 @@ import ClientsComponent from '../clients.component';
     MatFormFieldModule,
     DatePipe,
     DialogModule,
-    RouterLink
+    RouterLink,
+    NgIf
   ],
   templateUrl: './client-card.component.html',
   styleUrl: './client-card.component.scss',
   providers: [],
 })
-export class ClientCardComponent {
+export class ClientCardComponent extends BaseStoreComponent {
 
   @Input({ required: true }) client!: User;
 
@@ -46,9 +49,15 @@ export class ClientCardComponent {
 
   #router = inject(Router);
 
+  userId = computed(() => this.store.selectSignal(selectUserInfo)()?.id);
+
   closeDialog = signal(true);
-  
+
   async remove() {
+    if (this.userId() === this.client.id) {
+      this.#toastrService.warning('Operação não permitida', 'Aviso');
+      return;
+    }
     if (await this.#clientService.deleteById(this.client.id)) {
       this.#toastrService.success('Foi removido', 'Cliente');
       await this.#clientComponent.reload();
